Cleaned up more warnings: 3.

+++ src/levels.cpp
@@ -192,7 +192,7 @@ void level_class::load(std::string file_name)
             getline(level_file,data_line);
             {
                 temp_char = data_line[0];
-                if((temp_char != '#') && (data_line.length() > 2))
+                if((temp_char != '#') && ((int)data_line.length() > 2))
                 {
                     temp_char         = '#';
                     temp_string_key   = "";
@@ -203,13 +203,13 @@ void level_class::load(std::string file_name)
                         temp_char = data_line[count];
                         if(temp_char != ' ') temp_string_key += temp_char;
                         count++;
-                        if(count > data_line.length()) (temp_char = ' ');
+                        if(count > (int)data_line.length()) (temp_char = ' ');
                     }
                     while((temp_char == ' ') || (temp_char == '='))
                     {
                         temp_char = data_line[count];
                         count++;
-                        if(count > data_line.length()) (temp_char = '#');
+                        if(count > (int)data_line.length()) (temp_char = '#');
                     }
                     count--;
                     while(temp_char != ' ')
@@ -217,7 +217,7 @@ void level_class::load(std::string file_name)
                         temp_char = data_line[count];
                         if(temp_char != ' ') temp_string_value += temp_char;
                         count++;
-                        if(count > data_line.length()) (temp_char = ' ');
+                        if(count > (int)data_line.length()) (temp_char = ' ');
                     }
                     temp_float = atof(temp_string_value.c_str());
                     if (temp_string_key == "NPC_Type")
